cdk/pyproject.toml
[project]
name = "cdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aws-cdk-lib>=2.235.0",
    "constructor>=0.1.0",
]
